+++ src/classes/tournaments/tournament_champion.py
@@ -1,0 +1,5 @@
+class TournamentChampion():
+    def __init__(self, tournamentId, displayName, playerId):
+        self.tournamentId = tournamentId
+        self.displayName = displayName
+        self.playerId = playerId
